requirements.txt
numpy>=1.24.0
PyOpenGL>=3.1.6
PyOpenGL-accelerate>=3.1.6
PyQt6>=6.4.0
PyQt6-Qt6>=6.4.0
PyQt6-sip>=13.6.0
scipy>=1.10.0
pytest>=7.3.1 